Improve all the things...
- Move a few things around.
- Add a createComment API.
- Add a shouldUpdateProps reconciler API.
- Better match React heuristics with regards to Error Boundaries(starts
traversing from one level up the tree).
- Sync handle the pattern of setState in constructor via
`this.setState`.
- Opt to lazy-ly generate the componentStack string for caught error
boundaries.
- Improve internal WeakMap polyfill to use a deterministic random
number as a hash instead of a constant.
- Vanilla invoke custom-element/web-component constructors and leave
the heuristic of generating unique localNames to renderer implementors.
- Add tests for related changes, 100% coverage banzai.
- Improve documentations, in anticipation for a release.

The current heuristics for server rendering AsyncGenerators is to play
it save and only send the last state as the payload, maybe a future
release might render in-betweens if rendering intermediate values
proves to be a better UX.

The current heuristics for ErrorBoundaries that catch errors thrown in
their tree when server-side rendering with renderToNodeStream is to
play it safe and not render the thrown component/tree.

TODO:

- Performance tests.
- Third party breakage tests(this release is intended to only break the
heuristics around ErrorBoundaries that did not match React 1-to-1).
- Figure out a codename for the release ;)