Avoid terser compression
Terser sometimes inlines hole function expressions in other functions
and calls them within the said function instead of inlining the
contents of the function. This is great if you only invoke the function
in question once but that doesn’t seem to be what it does.

Removing the compress option doesn’t affect output by an reasonable
amount since most of the byte savings are in mangling symbols and name
bindings.